updated solvers, postprocessors, input file processors, and model classes. Working prototype functionality

+++ cofe_toolbox/input_processors/BulkEntryForce.m
@@ -18,7 +18,7 @@ classdef BulkEntryForce < BulkEntry
             % Construct using entry field data input as cell array of char
             obj.sid = castInputField('FORCE','SID',entryFields{2},'uint32',NaN,1);
             obj.g = castInputField('FORCE','G',entryFields{3},'uint32',NaN,1);
-            obj.cid = castInputField('FORCE','CID',entryFields{4},'uint32',0);            
+            obj.cid = castInputField('FORCE','CID',entryFields{4},'uint32',uint32(0));            
             obj.f = castInputField('FORCE','F',entryFields{5},'double',NaN);
             obj.n1 = castInputField('FORCE','N1',entryFields{6},'double',0.0);
             obj.n2 = castInputField('FORCE','N2',entryFields{7},'double',0.0);
